[#4] Extracted zoom functionality from CrafttechclientClient to Zoom
